refactor: pulled out calculate stats
